WebUI: convert to polymer 3 and es modules
- Move everything to es-modules rather than html imports. There are a number of ways these can be generated: completely manual or automatic via `polymer_modulizer` or `js_modulizer`. Template html can then optionally be 'included' in to the JS file via `html_to_js`.
- With a single patch in to a module requested by all webui pages (web_ui_listener_behavior.js), brave's polymer_overriding.js can inject any neccessary global template modifications.
