+++ browser/resources/settings/brave_appearance_page/super_referral.html
@@ -1,5 +1,6 @@
 <link rel="import" href="chrome://resources/html/polymer.html">
 <link rel="import" href="chrome://resources/html/i18n_behavior.html">
+<link rel="import" href="chrome://resources/html/assert.html">
 
 <link rel="import" href="chrome://resources/cr_elements/md_select_css.html">
