+++ browser/resources/settings/brave_clear_browsing_data_dialog/BUILD.gn
@@ -3,31 +3,16 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 
-import("//third_party/closure_compiler/compile_js.gni")
-import("//tools/polymer/polymer.gni")
-import("//ui/webui/resources/tools/js_modulizer.gni")
-import("../settings.gni")
-
-# ES Module conversion from HTML Imports code
+import("//tools/polymer/html_to_js.gni")
 
 group("web_modules") {
   public_deps = [
-    ":brave_clear_browsing_data_on_exit_page_module",
-    ":modules",
+    ":templatize",
   ]
 }
 
-polymer_modulizer("brave_clear_browsing_data_on_exit_page") {
-  js_file = "brave_clear_browsing_data_on_exit_page.js"
-  html_file = "brave_clear_browsing_data_on_exit_page.html"
-  html_type = "dom-module"
-  auto_imports = settings_auto_imports
-  namespace_rewrites = settings_namespace_rewrites
-}
-
-js_modulizer("modules") {
-  input_files = [
-    "brave_clear_browsing_data_dialog_behavior.js",
+html_to_js("templatize") {
+  js_files = [
+    "brave_clear_browsing_data_on_exit_page.js",
   ]
-  namespace_rewrites = settings_namespace_rewrites
 }
