--- browser/resources/settings/brave_clear_browsing_data_dialog/brave_clear_browsing_data_dialog_behavior.html
@@ -1,3 +1,0 @@
-<link rel="import" href="brave_clear_browsing_data_on_exit_page.html">
-<link rel="import" href="chrome://resources/html/i18n_behavior.html">
-<script src="brave_clear_browsing_data_dialog_behavior.js"></script>
