+++ browser/resources/settings/brave_clear_browsing_data_dialog/brave_clear_browsing_data_on_exit_page.html
@@ -1,76 +1,64 @@
-<link rel="import" href="chrome://resources/html/polymer.html">
+<style include="settings-shared">
 
-<link rel="import" href="chrome://resources/cr_elements/shared_vars_css.html">
-<link rel="import" href="chrome://resources/html/i18n_behavior.html">
-<link rel="import" href="../controls/settings_checkbox.html">
-<link rel="import" href="../settings_shared_css.html">
-<link rel="import" href="../settings_vars_css.html">
-
-<dom-module id="settings-brave-clear-browsing-data-on-exit-page">
-  <template>
-    <style include="settings-shared">
-
-      settings-checkbox {
-        --settings-row-two-line-min-height: 48px;
-        --settings-checkbox-label: {
-          line-height: 1.25rem;
-        }
-      }
-    </style>
-    <div id="checkboxes">
-      <settings-checkbox id="browsingCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.browsing_history_on_exit}}"
-                         label="$i18n{clearBrowsingHistory}"
-                         sub-label="[[counters.browsing_history]]"
-                         hidden="[[isSupervised_]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="downloadCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.download_history_on_exit}}"
-                         label="$i18n{clearDownloadHistory}"
-                         sub-label="[[counters.download_history]]"
-                         hidden="[[isSupervised_]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="cookiesCheckboxOnExit"
-                         class="cookies-checkbox"
-                         pref="{{prefs.browser.clear_data.cookies_on_exit}}"
-                         label="$i18n{clearCookies}"
-                         sub-label="[[counters.cookies]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="cacheCheckboxOnExit"
-                         class="cache-checkbox"
-                         pref="{{prefs.browser.clear_data.cache_on_exit}}"
-                         label="$i18n{clearCache}"
-                         sub-label="[[counters.cache]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="passwordsCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.passwords_on_exit}}"
-                         label="$i18n{clearPasswords}"
-                         sub-label="[[counters.passwords]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="formDataCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.form_data_on_exit}}"
-                         label="$i18n{clearFormData}"
-                         sub-label="[[counters.form_data]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="siteSettingsCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.site_settings_on_exit}}"
-                         label="$i18nPolymer{siteSettings}"
-                         sub-label="[[counters.site_settings]]"
-                         no-set-pref>
-      </settings-checkbox>
-      <settings-checkbox id="hostedAppDataCheckboxOnExit"
-                         pref="{{prefs.browser.clear_data.hosted_apps_data_on_exit}}"
-                         label="$i18n{clearHostedAppData}"
-                         sub-label="[[counters.hosted_apps_data]]"
-                         no-set-pref>
-      </settings-checkbox>
-    </div>
-  </template>
-  <script src="brave_clear_browsing_data_on_exit_page.js"></script>
-</dom-module>
+  settings-checkbox {
+    --settings-row-two-line-min-height: 48px;
+    --settings-checkbox-label: {
+      line-height: 1.25rem;
+    }
+  }
+</style>
+<div id="checkboxes">
+  <settings-checkbox id="browsingCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.browsing_history_on_exit}}"
+                      label="$i18n{clearBrowsingHistory}"
+                      sub-label="[[counters.browsing_history]]"
+                      hidden="[[isSupervised_]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="downloadCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.download_history_on_exit}}"
+                      label="$i18n{clearDownloadHistory}"
+                      sub-label="[[counters.download_history]]"
+                      hidden="[[isSupervised_]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="cookiesCheckboxOnExit"
+                      class="cookies-checkbox"
+                      pref="{{prefs.browser.clear_data.cookies_on_exit}}"
+                      label="$i18n{clearCookies}"
+                      sub-label="[[counters.cookies]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="cacheCheckboxOnExit"
+                      class="cache-checkbox"
+                      pref="{{prefs.browser.clear_data.cache_on_exit}}"
+                      label="$i18n{clearCache}"
+                      sub-label="[[counters.cache]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="passwordsCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.passwords_on_exit}}"
+                      label="$i18n{clearPasswords}"
+                      sub-label="[[counters.passwords]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="formDataCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.form_data_on_exit}}"
+                      label="$i18n{clearFormData}"
+                      sub-label="[[counters.form_data]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="siteSettingsCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.site_settings_on_exit}}"
+                      label="$i18nPolymer{siteSettings}"
+                      sub-label="[[counters.site_settings]]"
+                      no-set-pref>
+  </settings-checkbox>
+  <settings-checkbox id="hostedAppDataCheckboxOnExit"
+                      pref="{{prefs.browser.clear_data.hosted_apps_data_on_exit}}"
+                      label="$i18n{clearHostedAppData}"
+                      sub-label="[[counters.hosted_apps_data]]"
+                      no-set-pref>
+  </settings-checkbox>
+</div>
+</template>
