+++ browser/resources/settings/brave_icons.html
@@ -1,3 +1,4 @@
+<link rel="import" href="chrome://resources/html/polymer.html">
 <link rel="import"
   href="chrome://resources/polymer/v1_0/iron-iconset-svg/iron-iconset-svg.html">
 <!--
