+++ browser/resources/settings/brave_new_tab_page/BUILD.gn
@@ -3,7 +3,6 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # you can obtain one at http://mozilla.org/MPL/2.0/.
 
-import("//third_party/closure_compiler/compile_js.gni")
 import("//tools/polymer/polymer.gni")
 import("//ui/webui/resources/tools/js_modulizer.gni")
 import("../settings.gni")
@@ -12,12 +11,12 @@ import("../settings.gni")
 
 group("web_modules") {
   public_deps = [
-    ":brave_new_tab_page_module",
+    ":brave_new_tab_pages_module",
     ":modules",
   ]
 }
 
-polymer_modulizer("brave_new_tab_page") {
+polymer_modulizer("brave_new_tab_pages") {
   js_file = "brave_new_tab_page.js"
   html_file = "brave_new_tab_page.html"
   html_type = "dom-module"
