+++ browser/resources/settings/brave_new_tab_page/brave_new_tab_page.html
@@ -1,6 +1,6 @@
 <link rel="import" href="chrome://resources/html/polymer.html">
 
-<link rel="import" href="chrome://resources/html/i18n_behavior.html">
+<link rel="import" href="chrome://resources/html/web_ui_listener_behavior.html">
 <link rel="import" href="chrome://resources/cr_elements/md_select_css.html">
 <link rel="import" href="brave_new_tab_browser_proxy.html">
 <link rel="import" href="../settings_shared_css.html">
