+++ browser/resources/settings/brave_new_tab_page/brave_new_tab_page.js
@@ -1,8 +1,9 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
 
- (function() {
+(function() {
   'use strict';
 
   /**
@@ -48,6 +49,6 @@
       this.addWebUIListener('super-referral-active-state-changed', (isSuperReferralActive) => {
         this.isSuperReferralActive_ = isSuperReferralActive;
       })
-    },
+    }
   });
 })();
