+++ browser/resources/settings/brave_reset_page/brave_reset_profile_dialog_behavior.js
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-BraveResetProfileDialogBehavior = {
+export const BraveResetProfileDialogBehavior = {
 
   ready: function() {
     this.fixSendSettingsCheckbox();
