--- browser/resources/settings/brave_settings_overrides.html
@@ -1,215 +1,0 @@
-<link rel="import" href="brave_clear_browsing_data_dialog/brave_clear_browsing_data_dialog_behavior.html">
-<link rel="import" href="brave_reset_page/brave_reset_profile_dialog_behavior.html">
-<!-- Need chromium icons to be defined so that we can rewrite -->
-<link rel="import" href="icons.html">
-<link rel="import" href="brave_icons.html">
-<link rel="import" href="brave_appearance_page/super_referral.html">
-<link rel="import" href="brave_default_extensions_page/brave_default_extensions_page.html">
-<link rel="import" href="default_brave_shields_page/default_brave_shields_page.html">
-<link rel="import" href="social_blocking_page/social_blocking_page.html">
-<link rel="import" href="brave_sync_page/brave_sync_page.html">
-<link rel="import" href="brave_help_tips_page/brave_help_tips_page.html">
-<link rel="import" href="brave_new_tab_page/brave_new_tab_page.html">
-<link rel="import" href="site_settings/constants.html">
-
-<dom-module id="brave-settings-getting-started">
-  <template>
-    <style include="settings-shared md-select iron-flex">
-    </style>
-    <settings-people-page prefs="{{prefs}}"
-      page-visibility="[[pageVisibility]]">
-    </settings-people-page>
-    <settings-animated-pages id="pages" section="getStarted">
-      <div route-path="default">
-        <settings-default-browser-page></settings-default-browser-page>
-        <div class="settings-box">$i18n{onStartup}</div>
-        <div class="settings-box continuation">
-          <settings-on-startup-page prefs="{{prefs}}">
-          </settings-on-startup-page>
-        </div>
-      </div>
-    </settings-animated-pages>
-  </template>
-</dom-module>
-
-<!-- Style Overrides -->
-<dom-module id="brave-override-style-settings-ui">
-  <template>
-    <style>
-      :host {
-        display: inline-block;
-        /* we will enforce sizing at the container level and let children fill 100% of parent
-          top-down, rather than chromium's strategy of children setting their max width
-          bottom-up, which can be confusing, especially when our layout is different. */
-        --cr-centered-card-container_-_max-width: 100% !important; /* was 680px */
-        --brave-settings-content-max-width: 708px;
-        --brave-settings-menu-width: var(--settings-menu-width);
-        --brave-settings-menu-margin: 12px;
-      }
-      cr-drawer {
-        display: none !important;
-      }
-      #container {
-        /* menu and content next to each other in the horizontal center */
-        justify-content: center;
-        background: #F1F3F5; /* neutral100 */
-      }
-      #left {
-        /* fixed size menu */
-        flex: 0 0 calc(var(--settings-menu-width) + var(--brave-settings-menu-margin) * 2)  !important;
-      }
-      #main {
-        /* Take up rest of container up to a max */
-        flex: 0 1 var(--brave-settings-content-max-width) !important;
-      }
-      #right {
-        /* this element is only a space filler in chromium */
-        display: none;
-      }
-      @media (prefers-color-scheme: dark) {
-        #container {
-          background: #1E2127;
-        }
-      }
-    </style>
-  </template>
-</dom-module>
-
-<dom-module id="brave-override-style-settings-section">
-  <template>
-    <style include="settings-shared">
-
-    </style>
-  </template>
-</dom-module>
-
-<dom-module id="brave-override-style-settings-menu">
-  <template>
-    <style>
-      :host {
-        --brave-settings-menu-margin-v: 30px;
-        --brave-settings-menu-padding: 30px;
-        --settings-nav-item-color: #424242 !important;
-        position: sticky;
-        top: var(--brave-settings-menu-margin-v);
-        margin: 0 var(--brave-settings-menu-margin) !important;
-        max-height: calc(100vh - 56px - (var(--brave-settings-menu-margin-v) * 2) - (var(--brave-settings-menu-padding) * 2));
-        min-width: 172px;
-        border-radius: 6px;
-        background-color: #fff;
-        overflow-y: auto;
-        padding: 30px !important;
-      }
-
-      @media (prefers-color-scheme: dark) {
-        :host {
-          --settings-nav-item-color: #F4F4F4 !important;
-          border-color: transparent !important;
-          background-color: #161719;
-        }
-      }
-
-      a[href] {
-        font-weight: 400 !important;
-        margin: 0 20px 20px 0 !important;
-        margin-inline-start: 0 !important;
-        margin-inline-end: 0 !important;
-        padding-bottom: 0 !important;
-        padding-top: 0 !important;
-        padding-inline-start: 0 !important;
-      }
-
-      a[href].iron-selected {
-        color: var(--settings-nav-item-color) !important;
-        font-weight: 600 !important;
-      }
-
-      iron-icon {
-        color: var(--settings-nav-item-color);
-        margin-inline-end: 16px !important;
-        width: 20px;
-        height: 20px;
-      }
-
-      @media (prefers-color-scheme: dark) {
-        a[href].iron-selected iron-icon {
-          color: var(--settings-nav-item-color) !important;
-        }
-      }
-
-      a[href],
-      #advancedButton {
-        --cr-selectable-focus_-_outline: var(--brave-focus-outline) !important;
-      }
-
-      #advancedButton {
-        padding: 0 !important;
-        margin-top: 30px !important;
-        line-height: 1.25 !important;
-        border: none !important;
-      }
-
-      #advancedButton > iron-icon {
-        margin-inline-end: 0 !important;
-      }
-
-      #settingsHeader,
-      #advancedButton {
-        align-items: center !important;
-        font-weight: normal !important;
-        font-size: larger !important;
-        color: var(--settings-nav-item-color) !important;
-        margin-bottom: 20px !important;
-      }
-
-      #about-menu {
-        display: flex;
-        flex-direction: row;
-        align-items: flex-start;
-        justify-content: flex-start;
-        color: #c5c5d3 !important;
-        margin: 16px 0 0 0 !important;
-      }
-      .brave-about-graphic {
-        flex: 0;
-        flex-basis: 30%;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        align-self: stretch;
-      }
-      .brave-about-meta {
-        flex: 1;
-      }
-      .brave-about-item {
-        display: block;
-      }
-    </style>
-  </template>
-</dom-module>
-
-
-<dom-module id="brave-override-style-settings-basic-page">
-  <template>
-    <style>
-      :host {
-        min-width: 544px !important;
-      }
-
-    </style>
-  </template>
-</dom-module>
-
-<dom-module id="brave-override-style-settings-people-page">
-  <template>
-    <style>
-      #picture-subpage-trigger
-      {
-        display: none !important;
-      }
-    </style>
-  </template>
-</dom-module>
-
-<script src="brave_settings_overrides.js"></script>
-
