--- browser/resources/settings/brave_sync_page/brave_sync_browser_proxy.html
@@ -1,2 +1,0 @@
-<link rel="import" href="chrome://resources/html/cr.html">
-<script src="brave_sync_browser_proxy.js"></script>
