+++ browser/resources/settings/brave_sync_page/brave_sync_page.html
@@ -1,41 +1,22 @@
-<link rel="import" href="chrome://resources/html/polymer.html">
+<style include="settings-shared iron-flex">
+</style>
+<settings-animated-pages id="pages" section="braveSync">
+  <div route-path="default">
+    <cr-link-row id="brave-sync-setup"
+      label$="[[syncLabel_]]"
+      on-click="onSyncTap_"
+      role-description="$i18n{subpageArrowRoleDescription}">
+    </cr-link-row>
+  </div>
 
-<link rel="import" href="chrome://resources/html/i18n_behavior.html">
-<link rel="import" href="chrome://resources/html/web_ui_listener_behavior.html">
-<link rel="import" href="chrome://resources/cr_elements/icons.html">
-<link rel="import" href="../settings_page/settings_animated_pages.html">
-<link rel="import" href="../settings_page/settings_subpage.html">
-<link rel="import" href="../settings_shared_css.html">
-<link rel="import" href="../settings_vars_css.html">
-<link rel="import" href="brave_sync_subpage.html">
-<link rel="import" href="brave_sync_browser_proxy.html">
+  <template is="dom-if" route-path="/braveSync/setup">
+    <settings-subpage
+      associated-control="[[$$('#brave-sync-setup')]]"
+      page-title="[[syncLabel_]]">
+      <settings-brave-sync-subpage
+          sync-status="[[syncStatus_]]">
+      </settings-brave-sync-subpage>
+    </settings-subpage>
 
-<link rel="import" href="chrome://resources/cr_elements/cr_dialog/cr_dialog.html">
-
-<dom-module id="settings-brave-sync-page">
-  <template>
-    <style include="settings-shared iron-flex">
-    </style>
-    <settings-animated-pages id="pages" section="braveSync">
-    <div route-path="default">
-      <cr-link-row id="brave-sync-setup"
-        label$="[[syncLabel_]]"
-        on-click="onSyncTap_"
-        role-description="$i18n{subpageArrowRoleDescription}">
-      </cr-link-row>
-    </div>
-
-    <template is="dom-if" route-path="/braveSync/setup">
-      <settings-subpage
-        associated-control="[[$$('#brave-sync-setup')]]"
-        page-title="[[syncLabel_]]">
-        <settings-brave-sync-subpage
-            sync-status="[[syncStatus_]]">
-        </settings-brave-sync-subpage>
-      </settings-subpage>
-
-    </template>
-    </settings-animated-pages>
   </template>
-  <script src="brave_sync_page.js"></script>
-</dom-module>
+</settings-animated-pages>
