+++ browser/resources/settings/brave_sync_page/brave_sync_page.js
@@ -7,8 +7,26 @@
  * 'settings-brave-sync-page' is the settings page containing brave's
  * custom sync.
  */
+import 'chrome://resources/cr_elements/cr_dialog/cr_dialog.m.js';
+import 'chrome://resources/cr_elements/icons.m.js';
+import '../settings_page/settings_animated_pages.m.js';
+import '../settings_page/settings_subpage.m.js';
+import '../settings_shared_css.m.js';
+import '../settings_vars_css.m.js';
+import './brave_sync_subpage.js';
+
+import {I18nBehavior} from 'chrome://resources/js/i18n_behavior.m.js';
+import {WebUIListenerBehavior} from 'chrome://resources/js/web_ui_listener_behavior.m.js';
+import {Polymer, html} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js';
+
+import {Router} from '../router.m.js';
+import {SyncBrowserProxyImpl} from '../people_page/sync_browser_proxy.m.js';
+import {BraveSyncBrowserProxy} from './brave_sync_browser_proxy.js';
+
 Polymer({
   is: 'settings-brave-sync-page',
+
+  _template: html`{__html_template__}`,
 
   behaviors: [
     I18nBehavior,
@@ -18,7 +36,7 @@ Polymer({
   properties: {
     /**
      * The current sync status, supplied by SyncBrowserProxy.
-     * @type {?settings.SyncStatus}
+     * @type {?SyncStatus}
      */
     syncStatus_: Object,
     syncLabel_: {
@@ -27,15 +45,15 @@ Polymer({
     },
   },
 
-  /** @private {?settings.SyncBrowserProxy} */
+  /** @private {?SyncBrowserProxy} */
   browserProxy_: null,
   /** @private */
   braveBrowserProxy_: null,
 
   /** @override */
   created: function() {
-    this.browserProxy_ = settings.SyncBrowserProxyImpl.getInstance();
-    this.braveBrowserProxy_ = settings.BraveSyncBrowserProxy.getInstance();
+    this.browserProxy_ = SyncBrowserProxyImpl.getInstance();
+    this.braveBrowserProxy_ = BraveSyncBrowserProxy.getInstance();
   },
 
   /** @private */
@@ -58,13 +76,13 @@ Polymer({
   /** @private */
   onSyncTap_: function() {
     // Users can go to sync subpage regardless of sync status.
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     router.navigateTo(router.getRoutes().BRAVE_SYNC_SETUP);
   },
 
   /**
    * Handler for when the sync state is pushed from the browser.
-   * @param {?settings.SyncStatus} syncStatus
+   * @param {?SyncStatus} syncStatus
    * @private
    */
   handleSyncStatus_: function(syncStatus) {
