+++ browser/resources/settings/brave_sync_page/brave_sync_setup.html
@@ -1,78 +1,68 @@
-<link rel="import" href="chrome://resources/html/polymer.html">
 
-<link rel="import" href="brave_sync_browser_proxy.html">
-<link rel="import" href="brave_sync_code_dialog.html">
-
-<dom-module id="settings-brave-sync-setup">
-  <template>
-    <style include="settings-shared">
-      .sync-setup
-      {
-        /* break out of indent for the setup only */
-        margin-left: calc(var(--cr-section-indent-padding) * -1);
-        display: grid;
-        grid-template-columns: 30% auto;
-        grid-template-rows: auto;
-        column-gap: 20px;
-      }
-      .graphic img
-      {
-        max-width: 100%;
-      }
-      .actions
-      {
-        display: flex;
-        flex-direction: row;
-      }
-      .actions > *
-      {
-        margin-right: 8px;
-      }
-      .actions:last-child
-      {
-        margin-right: 0;
-      }
-      paper-spinner-lite
-      {
-        width: 16px;
-        height: 16px;
-      }
-    </style>
-    <div class="sync-setup">
-      <div class="graphic">
-        <img alt="" src="chrome://settings/brave_sync_page/start_icon.svg" />
-      </div>
-      <div class="content">
-        <h1>$i18n{braveSyncSetupTitle}</h1>
-        <p>$i18n{braveSyncSetupSubtitle}</p>
-        <div class="actions">
-          <cr-button class="action-button"
-            on-click="handleStartSyncChain_"
-            disabled="[[isGettingSyncCode_]]"
-          >
-            <paper-spinner-lite hidden="[[!isGettingSyncCode_]]">
-            </paper-spinner-lite>
-            <span hidden="[[isGettingSyncCode_]]">$i18n{braveSyncStartNewChain}</span>
-          </cr-button>
-          <cr-button
-            on-click="handleJoinSyncChain_"
-            disabled="[[isGettingSyncCode_]]"
-          >
-            $i18n{braveSyncEnterCode}
-          </cr-button>
-        </div>
-      </div>
+<style include="settings-shared">
+  .sync-setup
+  {
+    /* break out of indent for the setup only */
+    margin-left: calc(var(--cr-section-indent-padding) * -1);
+    display: grid;
+    grid-template-columns: 30% auto;
+    grid-template-rows: auto;
+    column-gap: 20px;
+  }
+  .graphic img
+  {
+    max-width: 100%;
+  }
+  .actions
+  {
+    display: flex;
+    flex-direction: row;
+  }
+  .actions > *
+  {
+    margin-right: 8px;
+  }
+  .actions:last-child
+  {
+    margin-right: 0;
+  }
+  paper-spinner-lite
+  {
+    width: 16px;
+    height: 16px;
+  }
+</style>
+<div class="sync-setup">
+  <div class="graphic">
+    <img alt="" src="chrome://settings/brave_sync_page/start_icon.svg" />
+  </div>
+  <div class="content">
+    <h1>$i18n{braveSyncSetupTitle}</h1>
+    <p>$i18n{braveSyncSetupSubtitle}</p>
+    <div class="actions">
+      <cr-button class="action-button"
+        on-click="handleStartSyncChain_"
+        disabled="[[isGettingSyncCode_]]"
+      >
+        <paper-spinner-lite hidden="[[!isGettingSyncCode_]]">
+        </paper-spinner-lite>
+        <span hidden="[[isGettingSyncCode_]]">$i18n{braveSyncStartNewChain}</span>
+      </cr-button>
+      <cr-button
+        on-click="handleJoinSyncChain_"
+        disabled="[[isGettingSyncCode_]]"
+      >
+        $i18n{braveSyncEnterCode}
+      </cr-button>
     </div>
+  </div>
+</div>
 
-    <template is="dom-if" if="[[syncCodeDialogType_]]" restamp>
-      <settings-brave-sync-code-dialog
-        code-type="{{syncCodeDialogType_}}"
-        sync-code="{{syncCode}}"
-        is-invalid-sync-code="{{isInvalidSyncCode_}}"
-        on-done="handleSyncCodeDialogDone_"
-      ></settings-brave-sync-code-dialog>
-    </template>
-
-  </template>
-  <script src="brave_sync_setup.js"></script>
-</dom-module>
+<template is="dom-if" if="[[syncCodeDialogType_]]" restamp>
+  <settings-brave-sync-code-dialog
+    code-type="{{syncCodeDialogType_}}"
+    sync-code="{{syncCode}}"
+    is-invalid-sync-code="{{isInvalidSyncCode_}}"
+    on-done="handleSyncCodeDialogDone_"
+  ></settings-brave-sync-code-dialog>
+</template>
