+++ browser/resources/settings/brave_sync_page/brave_sync_subpage.js
@@ -1,48 +1,62 @@
-/* Copyright (c) 2020 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
 
-(function() {
+
+import 'chrome://resources/js/util.m.js';
+
+import 'chrome://resources/cr_elements/cr_button/cr_button.m.js';
+import 'chrome://resources/cr_elements/cr_icon_button/cr_icon_button.m.js';
+import 'chrome://resources/cr_elements/icons.m.js';
+import 'chrome://resources/cr_elements/shared_style_css.m.js';
+import 'chrome://resources/cr_elements/shared_vars_css.m.js';
+import '../settings_shared_css.m.js';
+import '../settings_vars_css.m.js';
+import '../people_page/sync_controls.m.js';
+import './brave_sync_configure.js';
+import './brave_sync_setup.js';
+
+import {assert} from 'chrome://resources/js/assert.m.js';
+import {Polymer, html} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js';
+import {I18nBehavior} from 'chrome://resources/js/i18n_behavior.m.js';
+
+import {Route, RouteObserverBehavior, Router} from '../router.m.js';
+import {SyncBrowserProxyImpl, StatusAction} from '../people_page/sync_browser_proxy.m.js';
 
 /**
- * @fileoverview
- * 'settings-sync-page' is the settings page containing sync settings.
- */
+* @fileoverview
+* 'settings-sync-subpage' is the settings page content
+*/
 Polymer({
   is: 'settings-brave-sync-subpage',
 
+  _template: html`{__html_template__}`,
+
   behaviors: [
-    settings.RouteObserverBehavior,
+    RouteObserverBehavior,
   ],
 
   properties: {
-    /** @private */
-    pages_: {
-      type: Object,
-      value: settings.PageStatus,
-      readOnly: true,
-    },
-
     /**
-     * Current page status
-     * 'configure' | 'setup' | 'spinner'
-     * @private
-     */
+    * Current page status
+    * 'configure' | 'setup' | 'spinner'
+    * @private
+    */
     pageStatus_: {
       type: String,
       value: 'configure',
     },
 
     /**
-     * The current sync preferences, supplied by SyncBrowserProxy.
-     * @type {settings.SyncPrefs|undefined}
-     */
+    * The current sync preferences, supplied by SyncBrowserProxy.
+    * @type {SyncPrefs|undefined}
+    */
     syncPrefs: {
       type: Object,
     },
 
-    /** @type {settings.SyncStatus} */
+    /** @type {SyncStatus} */
     syncStatus: {
       type: Object,
     },
@@ -69,46 +83,46 @@ Polymer({
     'updatePageStatus_(syncStatus.*)'
   ],
 
-  /** @private {?settings.SyncBrowserProxy} */
+  /** @private {?SyncBrowserProxy} */
   browserProxy_: null,
 
   /**
-   * The beforeunload callback is used to show the 'Leave site' dialog. This
-   * makes sure that the user has the chance to go back and confirm the sync
-   * opt-in before leaving.
-   *
-   * This property is non-null if the user is currently navigated on the sync
-   * settings route.
-   *
-   * @private {?Function}
-   */
+  * The beforeunload callback is used to show the 'Leave site' dialog. This
+  * makes sure that the user has the chance to go back and confirm the sync
+  * opt-in before leaving.
+  *
+  * This property is non-null if the user is currently navigated on the sync
+  * settings route.
+  *
+  * @private {?Function}
+  */
   beforeunloadCallback_: null,
 
   /**
-   * The unload callback is used to cancel the sync setup when the user hits
-   * the browser back button after arriving on the page.
-   * Note: Cases like closing the tab or reloading don't need to be handled,
-   * because they are already caught in |PeopleHandler::~PeopleHandler|
-   * from the C++ code.
-   *
-   * @private {?Function}
-   */
+  * The unload callback is used to cancel the sync setup when the user hits
+  * the browser back button after arriving on the page.
+  * Note: Cases like closing the tab or reloading don't need to be handled,
+  * because they are already caught in |PeopleHandler::~PeopleHandler|
+  * from the C++ code.
+  *
+  * @private {?Function}
+  */
   unloadCallback_: null,
 
   /**
-   * Whether the user completed setup successfully.
-   * @private {boolean}
-   */
+  * Whether the user completed setup successfully.
+  * @private {boolean}
+  */
   setupSuccessful_: false,
 
   /** @override */
   created: function() {
-    this.browserProxy_ = settings.SyncBrowserProxyImpl.getInstance();
+    this.browserProxy_ = SyncBrowserProxyImpl.getInstance();
   },
 
   /** @override */
   attached: function() {
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     if (router.getCurrentRoute() == router.getRoutes().BRAVE_SYNC_SETUP) {
       this.onNavigateToPage_();
     }
@@ -116,7 +130,7 @@ Polymer({
 
   /** @override */
   detached: function() {
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     if (router.getRoutes().BRAVE_SYNC_SETUP.contains(router.getCurrentRoute())) {
       this.onNavigateAwayFromPage_();
     }
@@ -137,30 +151,30 @@ Polymer({
   },
 
   /**
-   * @return {boolean}
-   * @private
-   */
+  * @return {boolean}
+  * @private
+  */
   computeSyncSectionDisabled_: function() {
     return this.syncStatus !== undefined &&
         (!!this.syncStatus.disabled ||
-         (!!this.syncStatus.hasError &&
+        (!!this.syncStatus.hasError &&
           this.syncStatus.statusAction !==
-              settings.StatusAction.ENTER_PASSPHRASE &&
+              StatusAction.ENTER_PASSPHRASE &&
           this.syncStatus.statusAction !==
-              settings.StatusAction.RETRIEVE_TRUSTED_VAULT_KEYS));
+              StatusAction.RETRIEVE_TRUSTED_VAULT_KEYS));
   },
 
   /**
-   * @return {boolean}
-   * @private
-   */
+  * @return {boolean}
+  * @private
+  */
   computeSyncDisabledByAdmin_() {
     return this.syncStatus != undefined && !!this.syncStatus.managed;
   },
 
   /** @protected */
   currentRouteChanged: function() {
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     if (router.getCurrentRoute() == router.getRoutes().BRAVE_SYNC_SETUP) {
       this.onNavigateToPage_();
       return;
@@ -174,17 +188,17 @@ Polymer({
   },
 
   /**
-   * @param {!settings.PageStatus} expectedPageStatus
-   * @return {boolean}
-   * @private
-   */
+  * @param {!PageStatus} expectedPageStatus
+  * @return {boolean}
+  * @private
+  */
   isStatus_: function(expectedPageStatus) {
     return expectedPageStatus == this.pageStatus_;
   },
 
   /** @private */
   onNavigateToPage_: function() {
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     assert(router.getCurrentRoute() == router.getRoutes().BRAVE_SYNC_SETUP);
     if (this.beforeunloadCallback_) {
       return;
@@ -232,18 +246,16 @@ Polymer({
   },
 
   /**
-   * Called when setup is complete and sync code is set
-   * @private
-   */
+  * Called when setup is complete and sync code is set
+  * @private
+  */
   onSetupSuccess_: function() {
     this.setupSuccessful_ = true
     // This navigation causes the firstSetupInProgress flag to be marked as false
     // via `didNavigateAwayFromSyncPage`.
-    const router = settings.Router.getInstance();
+    const router = Router.getInstance();
     if (router.getCurrentRoute() == router.getRoutes().BRAVE_SYNC_SETUP) {
       router.navigateTo(router.getRoutes().BRAVE_SYNC);
     }
   },
 });
-
-})();
