+++ browser/resources/settings/getting_started_page/getting_started.js
@@ -1,0 +1,17 @@
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import {html, Polymer} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js'
+
+import '../people_page/people_page.js'
+import '../settings_page/settings_animated_pages.m.js'
+import '../default_browser_page/default_browser_page.js'
+import '../on_startup_page/on_startup_page.js'
+
+Polymer({
+  is: 'brave-settings-getting-started',
+
+  _template: html`{__html_template__}`
+})
