+++ browser/resources/settings/settings.gni
@@ -7,4 +7,9 @@ settings_namespace_rewrites += [
   "settings.BravePrivacyBrowserProxyImpl|BravePrivacyBrowserProxyImpl",
   "settings.BraveSyncBrowserProxy|BraveSyncBrowserProxy",
   "settings.DefaultBraveShieldsBrowserProxyImpl|DefaultBraveShieldsBrowserProxyImpl",
+  "settings.Router|Router"
+]
+
+settings_auto_imports += [
+  "chrome/browser/resources/settings/router.html|Router,Route,RouteObserverBehavior",
 ]
