+++ browser/resources/settings/settings_resources_generated.grdp
@@ -31,8 +31,8 @@
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_CLEAR_BROWSING_DATA_ON_EXIT_PAGE_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_clear_browsing_data_dialog/brave_clear_browsing_data_on_exit_page.m.js"
+  <include name="IDR_SETTINGS_BRAVE_CLEAR_BROWSING_DATA_ON_EXIT_PAGE_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_clear_browsing_data_dialog/brave_clear_browsing_data_on_exit_page.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
@@ -79,38 +79,32 @@
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_BROWSER_PROXY_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_browser_proxy.m.js"
+  <include name="IDR_SETTINGS_BRAVE_SYNC_CODE_DIALOG_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_code_dialog.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_CODE_DIALOG_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_code_dialog.m.js"
+  <include name="IDR_SETTINGS_BRAVE_SYNC_CONFIGURE_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_configure.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_CONFIGURE_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_configure.m.js"
+  <include name="IDR_SETTINGS_BRAVE_SYNC_PAGE_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_page.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_PAGE_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_page.m.js"
+  <include name="IDR_SETTINGS_BRAVE_SYNC_SETUP_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_setup.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
           type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_SETUP_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_setup.m.js"
-          use_base_dir="false"
-          preprocess="true"
-          compress="false"
-          type="BINDATA" />
-  <include name="IDR_SETTINGS_BRAVE_SYNC_SUBPAGE_M_JS"
-          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_subpage.m.js"
+  <include name="IDR_SETTINGS_BRAVE_SYNC_SUBPAGE_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_sync_page/brave_sync_subpage.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
@@ -127,8 +121,20 @@
           preprocess="true"
           compress="false"
           type="BINDATA" />
+  <include name="IDR_SETTINGS_BRAVE_GETTING_STARTED_PAGE_M_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/getting_started_page/getting_started.js"
+          use_base_dir="false"
+          preprocess="true"
+          compress="false"
+          type="BINDATA" />
   <include name="IDR_SETTINGS_SOCIAL_BLOCKING_PAGE_M_JS"
           file="${root_gen_dir}/brave/browser/resources/settings/social_blocking_page/social_blocking_page.m.js"
+          use_base_dir="false"
+          preprocess="true"
+          compress="false"
+          type="BINDATA" />
+  <include name="IDR_SETTINGS_BRAVE_ICONS_M_JS"
+          file="${root_gen_dir}/brave/browser/resources/settings/brave_icons.m.js"
           use_base_dir="false"
           preprocess="true"
           compress="false"
