+++ browser/resources/settings/social_blocking_page/social_blocking_page.js
@@ -2,9 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-(function() {
-'use strict';
-
 /**
  * 'settings-social-blocking-page' is the settings page containing brave's
  * social blocking options
@@ -12,4 +9,3 @@
 Polymer({
   is: 'settings-social-blocking-page'
 });
-})();
