+++ browser/resources/settings/style_overrides/index.js
@@ -1,0 +1,10 @@
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import './settings_basic_page.js'
+import './settings_menu.js'
+import './settings_people_page.js'
+import './settings_section.js'
+import './settings_ui.js'
