--- patches/chrome-browser-resources-settings-settings.html.patch
@@ -1,20 +1,0 @@
-diff --git a/chrome/browser/resources/settings/settings.html b/chrome/browser/resources/settings/settings.html
-index 8353825c6db5ef3636eebf54cacf8d8b9f2e29cc..3ef069431f09690636650408ed793dc8b502c482 100644
---- a/chrome/browser/resources/settings/settings.html
-+++ b/chrome/browser/resources/settings/settings.html
-@@ -8,6 +8,7 @@
-   <base href="chrome://settings">
- </if>
-   <link rel="stylesheet" href="chrome://resources/css/md_colors.css">
-+  <link rel="stylesheet" href="chrome://brave-resources/page_specific/settings/settings_loading_shim.css">
-   <style>
-     html {
-       background: var(--md-background-color);
-@@ -36,6 +37,7 @@
-   <settings-ui></settings-ui>
-   <link rel="stylesheet" href="chrome://resources/css/text_defaults_md.css">
-   <link rel="import" href="chrome://resources/html/polymer.html">
-+  <link rel="import" href="brave_settings_overrides.html">
-   <link rel="import" href="settings_ui/settings_ui.html">
- </body>
- </html>
