+++ patches/chrome-browser-resources-settings-settings_v3.html.patch
@@ -1,0 +1,12 @@
+diff --git a/chrome/browser/resources/settings/settings_v3.html b/chrome/browser/resources/settings/settings_v3.html
+index 510ade38c54cf830e2b4177ae88a97e540b631c0..c0c3dfb7810c4044a6099628b97badeb094aa955 100644
+--- a/chrome/browser/resources/settings/settings_v3.html
++++ b/chrome/browser/resources/settings/settings_v3.html
+@@ -5,6 +5,7 @@
+   <meta charset="utf-8">
+   <title>$i18n{settings}</title>
+   <base href="chrome://settings">
++  <link rel="stylesheet" href="chrome://brave-resources/page_specific/settings/settings_loading_shim.css">
+   <link rel="stylesheet" href="chrome://resources/css/md_colors.css">
+   <style>
+     html {
