--- patches/ui-webui-resources-html-polymer.html.patch
@@ -1,8 +1,0 @@
-diff --git a/ui/webui/resources/html/polymer.html b/ui/webui/resources/html/polymer.html
-index 6639e7dd54148fdfd24dfeb11a51a02f67b5ab4c..177bc66157ba124c4da18c15767f77e8a12836d4 100644
---- a/ui/webui/resources/html/polymer.html
-+++ b/ui/webui/resources/html/polymer.html
-@@ -1,2 +1,3 @@
- <script src="../js/polymer_config.js"></script>
- <link rel="import" href="chrome://resources/polymer/v1_0/polymer/polymer.html">
-+<link rel="import" href="chrome://brave-resources/polymer_overriding.html">
