+++ ui/webui/resources/brave_webui_resources.grd
@@ -29,10 +29,12 @@
              file="css/md_colors.css" type="chrome_html" compress="gzip" />
       <structure name="IDR_BRAVE_WEBUI_CSS_TEXT_DEFAULTS"
              file="css/text_defaults.css" type="chrome_html" compress="gzip" />
-      <structure name="IDR_BRAVE_WEBUI_POLYMER_OVERRIDING_HTML"
-             file="polymer_overriding.html" type="chrome_html" compress="gzip" />
       <structure name="IDR_BRAVE_WEBUI_POLYMER_OVERRIDING_JS"
              file="polymer_overriding.js" type="chrome_html" compress="gzip" />
+      <structure name="IDR_BRAVE_WEBUI_OVERRIDES_CR_TOGGLE_JS"
+             file="overrides/cr_toggle.js" type="chrome_html" compress="gzip" />
+      <structure name="IDR_BRAVE_WEBUI_OVERRIDES_CR_BUTTON_JS"
+             file="overrides/cr_button.js" type="chrome_html" compress="gzip" />
     </structures>
   </release>
 </grit>
