+++ ui/webui/resources/overrides/cr_toggle.js
@@ -1,0 +1,16 @@
+// Copyright (c) 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import {html} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js'
+
+export default html`
+  <style>
+    :host {
+      --cr-toggle-checked-bar-color: var(--cr-link-color) !important;
+      --cr-toggle-checked-button-color: var(--cr-link-color) !important;
+      --cr-toggle-checked-ink-color: var(--cr-link-color) !important;
+    }
+  </style>
+`
