+++ ui/webui/resources/polymer_overriding.js
@@ -3,123 +3,198 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at http://mozilla.org/MPL/2.0/.
 
-// This module is in global scope,
-// so create a temporary scope.
-(function () {
-  const debug = false
-  // Re-write `Polymer(componentInfo)` function
-  const oldFn = window.Polymer._polymerFn
-  function newPolymerFn(component) {
-    if (component && component.is) {
-      if (debug) {
-        console.debug(`Polymer component registering: ${component.is}`, component)
-      }
-      addBraveBehaviors(component)
-      addBraveTemplateModifications(component)
-      addBraveStyleOverrides(component)
-    }
-    return oldFn(component)
+import {Polymer, PolymerElement} from 'chrome://resources/polymer/v3_0/polymer/polymer_bundled.min.js';
+
+// Global overrides
+import CrButtonStyleTemplate from './overrides/cr_button.js'
+import CrToggleStyleTemplate from './overrides/cr_toggle.js'
+
+const debug = false
+
+if (debug) {
+  // Useful to diagnose module definition timing for template modification
+  // debugging.
+  console.log("MODULE: polymer_overriding")
+}
+
+const allBehaviorsMap = {}
+
+function addBraveBehaviors(moduleName, component) {
+  if (allBehaviorsMap[moduleName]) {
+    component.behaviors = component.behaviors || []
+    component.behaviors.push(...allBehaviorsMap[moduleName])
+    delete allBehaviorsMap[moduleName]
   }
-  window.Polymer._polymerFn = newPolymerFn
+}
 
-  const allBehaviorsMap = {}
+const allBraveTemplateModificationsMap = {}
 
-  function addBraveBehaviors(component) {
-    if (allBehaviorsMap[component.is]) {
-      component.behaviors = component.behaviors || []
-      component.behaviors.push(...allBehaviorsMap[component.is])
-      delete allBehaviorsMap[component.is]
+function addBraveTemplateModifications(moduleName, component) {
+  if (allBraveTemplateModificationsMap[moduleName]) {
+    const template = component.template || component._template
+    if (template) {
+      const templateContent = template.content
+      const t0 = debug && performance.now()
+      allBraveTemplateModificationsMap[moduleName](templateContent)
+      const t1 = debug && performance.now()
+      if (debug)
+        console.debug(`Modifying template '${moduleName}' took ${t1 - t0}ms`)
+    } else {
+      console.error(`Source template not found for override of component "${moduleName}"`)
     }
+    delete allBraveTemplateModificationsMap[moduleName]
   }
+}
 
-  const allBraveTemplateModificationsMap = {}
+const styleOverridePrefix = 'brave-override-style-'
 
-  function addBraveTemplateModifications(component) {
-    if (allBraveTemplateModificationsMap[component.is]) {
-      const moduleName = component.is
-      const domModule = Polymer.DomModule.import(moduleName)
-      if (domModule) {
-        const template = domModule.querySelector('template')
-        if (template) {
-          const templateContent = template.content
-          const t0 = performance.now()
-          allBraveTemplateModificationsMap[component.is](templateContent)
-          const t1 = performance.now()
-          console.debug(`Modifying template '${component.is}' took ${t1 - t0}ms`)
-        }
-      }
-      delete allBraveTemplateModificationsMap[component.is]
-    }
+function addBraveStyleOverride(moduleName, component) {
+  // does have template style element?
+  const template = component.template || component._template
+  if (!template) {
+    console.error(`No template found for component (${moduleName}) with found style overrides`, component)
+    return
   }
-
-  function addBraveStyleOverrides(component) {
-    // does have template style element?
-    const moduleName = component.is
-    const domModule = Polymer.DomModule.import(moduleName)
-    if (!domModule) {
-      return
-    }
-    // has override?
-    const overrideModuleName = `brave-override-style-${moduleName}`
-    const hasOverride = !!Polymer.DomModule.import(overrideModuleName)
-    if (!hasOverride) {
-      return
-    }
-    const template = domModule.querySelector('template')
-    if (!template) {
-      return
-    }
-    const styleElement = template.content.querySelector('style')
-    if (!styleElement) {
-      return
-    }
-    styleElement.setAttribute(
-      'include',
-      `${styleElement.getAttribute('include')} ${overrideModuleName}`
-    )
+  const styleElement = template.content.querySelector('style')
+  if (!styleElement) {
+    console.error(`No style element found for component (${moduleName}) with found style overrides`, component)
+    return
   }
+  const overrideModuleName = styleOverridePrefix + moduleName
+  styleElement.setAttribute(
+    'include',
+    `${styleElement.getAttribute('include')} ${overrideModuleName}`
+  )
+  if (debug)
+    console.log(`Brave Style Override added for ${moduleName}`, styleElement)
+}
 
-  function RegisterPolymerComponentBehaviors(behaviorsMap) {
-    Object.assign(allBehaviorsMap, behaviorsMap)
+export function RegisterPolymerComponentBehaviors(behaviorsMap) {
+  if (debug) {
+    console.log('RegisterPolymerComponentBehaviors', Object.keys(behaviorsMap))
   }
+  Object.assign(allBehaviorsMap, behaviorsMap)
+}
 
-  function RegisterPolymerTemplateModifications(modificationsMap) {
-    Object.assign(allBraveTemplateModificationsMap, modificationsMap)
+export function RegisterPolymerTemplateModifications(modificationsMap) {
+  Object.assign(allBraveTemplateModificationsMap, modificationsMap)
+}
+
+const moduleNamesWithStyleOverrides = []
+
+export async function RegisterStyleOverride(componentName, styleTemplate) {
+  // Inform polymer of styles by creating a dom-module.
+  //   <dom-module id="${styleOverridePrefix + componentName}">
+  //     <template>
+  //       ${styleOuterHTML}
+  //     </template>
+  //   </dom-module>
+  // This is the same way that polymer_modulizer uses for style-module types.
+  const template = document.createElement('template')
+  const domModule = document.createElement('dom-module')
+  domModule.id = styleOverridePrefix + componentName
+  template.appendChild(domModule)
+  domModule.appendChild(styleTemplate)
+  document.body.appendChild(template.content.cloneNode(true))
+  if (debug) {
+    console.log(`REGISTERING STYLE OVERRIDE for ${componentName}`, styleTemplate)
   }
+  // If module has already been defined, it's not ideal as the unmodified
+  // template may already have been cloned for an element. However, let's make
+  // an attempt and apply it anyway. Otherwise, we wait until it's defined and
+  // then modify the template as soon as possible.
+  const existingComponent = window.customElements.get(componentName)
+  if (existingComponent) {
+    addBraveStyleOverride(componentName, existingComponent)
+  } else {
+    // Cannot await CustomElementRegistry.whenDefined here
+    // since getting in the async queue will mean this template
+    // mofification happens too late. Instead, save this in a list
+    // so that the template modification can happen inside the
+    // customElements.define hook.
+    moduleNamesWithStyleOverrides.push(componentName)
+  }
+}
 
-  function OverrideIronIcons(iconSetName, overridingIconSetName, iconOverrides) {
-    // Here, we change the underlying DOM.
-    // We cannot simply change `IconSet._icons` since
-    // there are occasions where this will get re-parsed from DOM elements.
-    const meta = Polymer.Base.create('iron-meta', {type: 'iconset'})
-    const srcIconSet = meta.byKey(iconSetName)
-    const overrideIconSet = meta.byKey(overridingIconSetName)
-    for (const chromiumIconName in iconOverrides) {
-      const chromiumIcon = srcIconSet.querySelector(`#${chromiumIconName}`)
-      if (!chromiumIcon) {
-        console.error(`[brave overrides] Could not find chromium icon '${chromiumIconName}' in iconset '${iconSetName}' for replacement!`)
-        continue
-      }
-      const braveIconName = iconOverrides[chromiumIconName]
-      const braveIcon = overrideIconSet.querySelector(`#${braveIconName}`)
-      if (!braveIcon) {
-        console.error(`[brave overrides] Could not find brave icon '${braveIconName}' in iconset '${overridingIconSetName}' for replacement!`)
-        continue
-      }
-      // replace
-      chromiumIcon.innerHTML = braveIcon.innerHTML
+export function OverrideIronIcons(iconSetName, overridingIconSetName, iconOverrides) {
+  // Here, we change the underlying DOM.
+  // We cannot simply change `IconSet._icons` since
+  // there are occasions where this will get re-parsed from DOM elements.
+  // const meta = Base.create('iron-meta', {type: 'iconset'})
+  const meta = document.createElement('iron-meta')
+  meta.type = 'iconset'
+  const srcIconSet = meta.byKey(iconSetName)
+  if (!srcIconSet) {
+    console.error(`OverrideIronIcons: source icon set "${iconSetName} not found.`)
+    return
+  }
+  const overrideIconSet = meta.byKey(overridingIconSetName)
+  if (!overrideIconSet) {
+    console.error(`OverrideIronIcons: overriding icon set "${overridingIconSetName} not found.`)
+    return
+  }
+  for (const chromiumIconName in iconOverrides) {
+    const chromiumIcon = srcIconSet.querySelector(`#${chromiumIconName}`)
+    if (!chromiumIcon) {
+      console.error(`[brave overrides] Could not find chromium icon '${chromiumIconName}' in iconset '${iconSetName}' for replacement!`)
+      continue
     }
-    // Ensure icons get re-parsed if already parseds
-    // `getIconNames` ensures this._icons in iron-iconset-svg is re-parsed
-    // from DOM. If that changes, we'll need to find another way,
-    // perhaps `srcIconSet._icons = srcIconSet._createIconMap()`
-    srcIconSet.getIconNames()
+    const braveIconName = iconOverrides[chromiumIconName]
+    const braveIcon = overrideIconSet.querySelector(`#${braveIconName}`)
+    if (!braveIcon) {
+      console.error(`[brave overrides] Could not find brave icon '${braveIconName}' in iconset '${overridingIconSetName}' for replacement!`)
+      continue
+    }
+    // replace
+    chromiumIcon.innerHTML = braveIcon.innerHTML
   }
+  // Ensure icons get re-parsed if already parseds
+  // `getIconNames` ensures this._icons in iron-iconset-svg is re-parsed
+  // from DOM. If that changes, we'll need to find another way,
+  // perhaps `srcIconSet._icons = srcIconSet._createIconMap()`
+  srcIconSet.getIconNames()
+}
 
-  // Accessible to other modules
-  window.BravePatching = Object.assign({}, window.BravePatching, {
-    RegisterPolymerComponentBehaviors,
-    RegisterPolymerTemplateModifications,
-    OverrideIronIcons
-  })
-})()
+function PerformBraveModifications(name, component) {
+  if (debug) {
+    console.debug(`Polymer component registering: ${name}`, component)
+  }
+  addBraveBehaviors(name, component)
+  addBraveTemplateModifications(name, component)
+  if (moduleNamesWithStyleOverrides.includes(name)) {
+    addBraveStyleOverride(name, component)
+  }
+}
+
+// TODO(petemill): Overriding Polymer.Class only works because
+// chromium components at the moment are passing objects rather
+// than classes. If this changes, or for something more robust,
+// we can instead hook in to `window.customElements.define`. This
+// will require changing how we inject behaviors to instead return a
+// subclass of the original component, with the lifecycle methods added.
+// That's because behaviors are a legacy polymer feature,
+// now migrated to subclassing.
+// That should work for any type of Polymer component (class or
+// object-to-generated-class).
+const oldClass = Polymer.Class
+Polymer.Class = function (info, mixin) {
+  if (!info) {
+    console.warn('Polymer.Class requires `info` argument');
+  }
+  const name = info.is
+  if (!name) {
+    if (debug) {
+      console.warn(`Non-polymer-like polymer class`, info)
+    }
+    return oldClass(info, mixin)
+  }
+  if (debug) {
+    console.log('defined', name)
+  }
+  PerformBraveModifications(name, info)
+  return oldClass(info, mixin)
+}
+
+// Overrides for all pages
+RegisterStyleOverride('cr-toggle', CrToggleStyleTemplate)
+RegisterStyleOverride('cr-button', CrButtonStyleTemplate)
